requirements.txt
torch>=2.0.0
transformers>=4.36.0
unsloth
trl>=0.7.0
accelerate>=0.25.0
peft>=0.7.0
datasets
pandas
tqdm
matplotlib
scikit-learn
Pillow
python-dotenv
rouge-score
evaluate
bert-score
nltk
openpyxl